Factor chat related code out of CandidateOpportunityServiceImpl into OppNotificationService
