+++ server/src/main/java/org/tctalent/server/service/db/impl/CandidateOpportunityServiceImpl.java
@@ -56,12 +56,8 @@ import org.tctalent.server.model.db.CandidateOpportunity;
 import org.tctalent.server.model.db.CandidateOpportunityStage;
 import org.tctalent.server.model.db.CandidateOpportunityStageHistory;
 import org.tctalent.server.model.db.CandidateStatus;
-import org.tctalent.server.model.db.ChatPost;
-import org.tctalent.server.model.db.JobChat;
-import org.tctalent.server.model.db.JobChatType;
 import org.tctalent.server.model.db.SalesforceJobOpp;
 import org.tctalent.server.model.db.User;
-import org.tctalent.server.model.db.chat.Post;
 import org.tctalent.server.model.db.partner.Partner;
 import org.tctalent.server.model.sf.Opportunity;
 import org.tctalent.server.model.sf.OpportunityHistory;
@@ -75,9 +71,8 @@ import org.tctalent.server.request.candidate.opportunity.SearchCandidateOpportun
 import org.tctalent.server.security.AuthService;
 import org.tctalent.server.service.db.CandidateOpportunityService;
 import org.tctalent.server.service.db.CandidateService;
-import org.tctalent.server.service.db.ChatPostService;
 import org.tctalent.server.service.db.FileSystemService;
-import org.tctalent.server.service.db.JobChatService;
+import org.tctalent.server.service.db.OppNotificationService;
 import org.tctalent.server.service.db.SalesforceJobOppService;
 import org.tctalent.server.service.db.SalesforceService;
 import org.tctalent.server.service.db.UserService;
@@ -90,16 +85,15 @@ import org.tctalent.server.util.filesystem.GoogleFileSystemFolder;
 @RequiredArgsConstructor
 @Slf4j
 public class CandidateOpportunityServiceImpl implements CandidateOpportunityService {
+    private final AuthService authService;
     private final CandidateOpportunityRepository candidateOpportunityRepository;
     private final CandidateService candidateService;
-    private final JobChatService jobChatService;
+    private final FileSystemService fileSystemService;
+    private final GoogleDriveConfig googleDriveConfig;
+    private final OppNotificationService oppNotificationService;
     private final SalesforceJobOppService salesforceJobOppService;
     private final SalesforceService salesforceService;
     private final UserService userService;
-    private final AuthService authService;
-    private final GoogleDriveConfig googleDriveConfig;
-    private final FileSystemService fileSystemService;
-    private final ChatPostService chatPostService;
 
     /**
      * Creates or updates CandidateOpportunities associated with the given candidates going for
@@ -155,10 +149,8 @@ public class CandidateOpportunityServiceImpl implements CandidateOpportunityServ
         opp = updateCandidateOpportunity(opp, oppParams);
 
         if (create) {
-            // todo why are we creating a recruiting chat here? Candidate is only now a prospect?
-            jobChatService.createCandidateRecruitingChat(opp.getCandidate(), opp.getJobOpp());
-            //Create the automated post to notify that a candidate has been added to a submission list.
-            publishAddedToSubmissionListPosts(opp);
+            //Create the automated post to notify that a new has been created for a candidate.
+            oppNotificationService.notifyNewOpp(opp);
         }
 
         return opp;
@@ -644,21 +636,13 @@ public class CandidateOpportunityServiceImpl implements CandidateOpportunityServ
         CandidateOpportunity opp, @Nullable CandidateOpportunityParams oppParams) {
 
         if (oppParams != null) {
+
+            oppNotificationService.notifyOppChanges(opp, oppParams);
+
             final CandidateOpportunityStage newStage = oppParams.getStage();
             if (newStage != null) {
                 // Stage is changing
                 if (!newStage.equals(opp.getStage())) {
-                    // If stage is changing to CLOSED (e.g. removed from submission list) publish posts
-                    if (newStage.isClosed()) {
-                        publishRemovedFromSubmissionListPosts(opp, newStage);
-                    // If a stage is changed to ACCEPTANCE (job offer is accepted)
-                    } else if (newStage.equals(CandidateOpportunityStage.acceptance)) {
-                        publishOppAcceptedPosts(opp);
-                    } else {
-                        // If non closing stage change, publish posts
-                        publishStageChangePosts(opp, newStage);
-                    }
-
                     // If new stage is relocated or above AND not closed or won, then set relocated address
                     if (newStage.compareTo(CandidateOpportunityStage.relocated) >= 0 && (!newStage.isClosed() || newStage.isWon())) {
                         updateCandidateRelocatedCountry(opp);
@@ -680,57 +664,6 @@ public class CandidateOpportunityServiceImpl implements CandidateOpportunityServ
                     opp.getNextStep(), oppParams.getNextStep());
 
                 opp.setNextStep(processedNextStep);
-
-                // If next step details changed, automate post to JobCreatorSourcePartner chat.
-                // To compare previous next step to new one, need to ensure neither is null.
-                // Cases are auto-populated with a value for next step when created, but this has
-                // not always been the case.
-                String currentNextStep = opp.getNextStep() == null ? "" : opp.getNextStep();
-
-                // If only the due date has changed, we still want to send a message.
-                // As above, there are some old cases with null values that need to be dealt with.
-                LocalDate currentDueDate =
-                    opp.getNextStepDueDate() == null ?
-                        LocalDate.of(1970, 1, 1) : opp.getNextStepDueDate();
-
-                // If the request due date is null (user deletes the existing value in the form but
-                // doesn't set a new one, then submits) it will not be used (see below) — so, for
-                // purpose of comparison we give it the same value as the current due date (no
-                // message will be sent because they're the same).
-                LocalDate requestDueDate =
-                    oppParams.getNextStepDueDate() == null ?
-                        currentDueDate : oppParams.getNextStepDueDate();
-
-                if (!processedNextStep.equals(currentNextStep) || !requestDueDate.equals(currentDueDate)) {
-                    // Find the relevant job chat
-                    JobChat jcspChat = jobChatService.getOrCreateJobChat(
-                        JobChatType.JobCreatorSourcePartner,
-                        opp.getJobOpp(),
-                        opp.getCandidate().getUser().getPartner(),
-                        null
-                    );
-
-                    String candidateNameAndNumber = constructCandidateNameNumber(opp.getCandidate());
-
-                    // Set the chat post content
-                    Post autoPostNextStepChange = new Post();
-                    autoPostNextStepChange.setContent(
-                        "💼 <b>" + opp.getName()
-                            + "</b> 🪜<br> The next step details have changed for this case relating to candidate "
-                            + candidateNameAndNumber
-                            + ".<br><b>Next step:</b> " + processedNextStep
-                            + "<br><b>Due date:</b> "
-                            + (oppParams.getNextStepDueDate() == null ?
-                            opp.getNextStepDueDate() : oppParams.getNextStepDueDate())
-                    );
-
-                    // Create the chat post
-                    ChatPost nextStepChangeChatPost = chatPostService.createPost(
-                        autoPostNextStepChange, jcspChat, userService.getSystemAdminUser());
-
-                    // Publish the chat post
-                    chatPostService.publishChatPost(nextStepChangeChatPost);
-                }
             }
 
             final LocalDate requestDueDate = oppParams.getNextStepDueDate();
@@ -826,169 +759,6 @@ public class CandidateOpportunityServiceImpl implements CandidateOpportunityServ
     }
 
     /**
-     * Publish posts for a candidate opportunity that is moved to a closing stage.
-     * Publish to:
-     * - JobCreatorSourcePartner chat
-     * - CandidateProspect chat
-     * - CandidateRecruiting chat
-     * @param opp CandidateOpportunity - the candidate opp that's stage is being changed
-     * @param newStage CandidateOpportunityStage - closing stage that the opp is being changed to
-     */
-    private void publishRemovedFromSubmissionListPosts(CandidateOpportunity opp, CandidateOpportunityStage newStage) {
-        Candidate candidate = opp.getCandidate();
-        String candidateNameAndNumber = constructCandidateNameNumber(opp.getCandidate());
-
-        Post autoPostRemovedFromSubList = new Post();
-        autoPostRemovedFromSubList.setContent("The candidate " + candidateNameAndNumber +
-                " has been removed for the job '" + opp.getJobOpp().getName() +
-                "' with the reason " + newStage.getSalesforceStageName() + ".");
-
-        //Only post to candidate if they got past the prospect stage for this job
-        if (opp.getStage() != CandidateOpportunityStage.prospect) {
-            // AUTO CHAT TO PROSPECT CHAT
-            JobChat prospectChat = jobChatService.getOrCreateJobChat(JobChatType.CandidateProspect, null,
-                null, candidate);
-            // Create the chat post
-            ChatPost prospectChatPostRemoved = chatPostService.createPost(
-                autoPostRemovedFromSubList, prospectChat, userService.getSystemAdminUser());
-            // Publish chat post
-            chatPostService.publishChatPost(prospectChatPostRemoved);
-        }
-
-        // AUTO CHAT TO RECRUITING CHAT
-        JobChat recruitingChat = jobChatService.getOrCreateJobChat(JobChatType.CandidateRecruiting, opp.getJobOpp(),
-                candidate.getUser().getPartner(), candidate);
-        // Create the chat post
-        ChatPost recruitingChatPostRemoved = chatPostService.createPost(
-                autoPostRemovedFromSubList, recruitingChat, userService.getSystemAdminUser());
-        // Publish chat post
-        chatPostService.publishChatPost(recruitingChatPostRemoved);
-
-        // AUTO CHAT TO JOB CREATOR SOURCE PARTNER CHAT
-        JobChat jcspChat = jobChatService.getOrCreateJobChat(JobChatType.JobCreatorSourcePartner, opp.getJobOpp(),
-                candidate.getUser().getPartner(), null);
-        // Create the chat post
-        ChatPost jcspChatPostRemoved = chatPostService.createPost(
-                autoPostRemovedFromSubList, jcspChat, userService.getSystemAdminUser());
-        // Publish chat post
-        chatPostService.publishChatPost(jcspChatPostRemoved);
-    }
-
-    /**
-     * Publish posts for after a Candidate Opportunity is created (e.g. Added to submission list)
-     * and the stage is now prospect.
-     * Publish to:
-     * - CandidateProspect chat
-     * - JobCreatorSourcePartner chat.
-     * @param opp CandidateOpportunity - the candidate opp that's stage is being changed
-     */
-    private void publishAddedToSubmissionListPosts(CandidateOpportunity opp) {
-        Candidate candidate = opp.getCandidate();
-        String candidateNameAndNumber = constructCandidateNameNumber(opp.getCandidate());
-
-        Post autoPostAddedToSubList = new Post();
-        autoPostAddedToSubList.setContent("The candidate " + candidateNameAndNumber +
-                " is a prospect for the job '" + opp.getJobOpp().getName() +"'.");
-
-        // Note that we don't post to candidates until they get past the prospect stage
-        if (opp.getStage() != CandidateOpportunityStage.prospect) {
-            // AUTO CHAT TO PROSPECT CHAT
-            JobChat prospectChat = jobChatService.getOrCreateJobChat(JobChatType.CandidateProspect,
-                null,null, candidate);
-            // Create the chat post
-            ChatPost prospectChatPost = chatPostService.createPost(
-                autoPostAddedToSubList, prospectChat, userService.getSystemAdminUser());
-            //publish chat post
-            chatPostService.publishChatPost(prospectChatPost);
-        }
-
-        // AUTO CHAT TO JOB CREATOR SOURCE PARTNER CHAT
-        JobChat jcspChat = jobChatService.getOrCreateJobChat(JobChatType.JobCreatorSourcePartner, opp.getJobOpp(),
-                candidate.getUser().getPartner(), null);
-        // Create the chat post
-        ChatPost jcspChatPost = chatPostService.createPost(
-                autoPostAddedToSubList, jcspChat, userService.getSystemAdminUser());
-        //publish chat post
-        chatPostService.publishChatPost(jcspChatPost);
-    }
-
-    /**
-     * Publish post for a candidate opportunity stage change that is not to a closing stage.
-     * Publish to Job Creator Source Partner chat.
-     * @param opp CandidateOpportunity - the candidate opp that's stage is being changed
-     * @param newStage CandidateOpportunityStage - non-closing stage that the opp is being changed to
-     */
-    private void publishStageChangePosts(CandidateOpportunity opp, CandidateOpportunityStage newStage) {
-        // Find the relevant job chat
-        JobChat jcspChat = jobChatService.getOrCreateJobChat(
-                JobChatType.JobCreatorSourcePartner,
-                opp.getJobOpp(),
-                opp.getCandidate().getUser().getPartner(),
-                null
-        );
-
-        String candidateNameAndNumber = constructCandidateNameNumber(opp.getCandidate());
-
-        // Set the chat post content
-        Post autoPostCandidateOppStageChange = new Post();
-        autoPostCandidateOppStageChange.setContent(
-                "💼 <b>" + opp.getName() + "</b> 🪜<br> This case for candidate "
-                        + candidateNameAndNumber
-                        + " has changed stage from '" + opp.getStage() + "' to '"
-                        + newStage + "'."
-        );
-
-        // Create the chat post
-        ChatPost candidateOppStageChangeChatPost = chatPostService.createPost(
-                autoPostCandidateOppStageChange, jcspChat, userService.getSystemAdminUser());
-
-        // Publish the chat post
-        chatPostService.publishChatPost(candidateOppStageChangeChatPost);
-    }
-
-    /**
-     * Publish post for a candidate opportunity stage change to acceptance. Notify all previous chats.
-     * - CandidateProspect chat
-     * - CandidateRecruiting chat
-     * - JobCreatorSourcePartner chat
-     * @param opp CandidateOpportunity - the candidate opp that's stage is being changed
-     */
-    private void publishOppAcceptedPosts(CandidateOpportunity opp) {
-        Candidate candidate = opp.getCandidate();
-        String candidateNameAndNumber = constructCandidateNameNumber(opp.getCandidate());
-        Post autoPostAcceptedJobOffer = new Post();
-        autoPostAcceptedJobOffer.setContent("The candidate " + candidateNameAndNumber + " has accepted the job offer from '"
-                + opp.getJobOpp().getName() + " and is now a member of the <a href=\"https://pathwayclub.org/about\" target=\"_blank\">Pathway Club</a>.");
-
-        // AUTO CHAT TO PROSPECT CHAT
-        JobChat prospectChat = jobChatService.getOrCreateJobChat(JobChatType.CandidateProspect, null,
-                null, candidate);
-        // Create the chat post
-        ChatPost prospectChatPostAccepted = chatPostService.createPost(
-                autoPostAcceptedJobOffer, prospectChat, userService.getSystemAdminUser());
-        // Publish chat post
-        chatPostService.publishChatPost(prospectChatPostAccepted);
-
-        // AUTO CHAT TO RECRUITING CHAT
-        JobChat recruitingChat = jobChatService.getOrCreateJobChat(JobChatType.CandidateRecruiting, opp.getJobOpp(),
-                candidate.getUser().getPartner(), candidate);
-        // Create the chat post
-        ChatPost recruitingChatPostAccepted = chatPostService.createPost(
-                autoPostAcceptedJobOffer, recruitingChat, userService.getSystemAdminUser());
-        // Publish chat post
-        chatPostService.publishChatPost(recruitingChatPostAccepted);
-
-        // AUTO CHAT TO JOB CREATOR SOURCE PARTNER CHAT
-        JobChat jcspChat = jobChatService.getOrCreateJobChat(JobChatType.JobCreatorSourcePartner, opp.getJobOpp(),
-                candidate.getUser().getPartner(), null);
-        // Create the chat post
-        ChatPost jcspChatPostAccepted = chatPostService.createPost(
-                autoPostAcceptedJobOffer, jcspChat, userService.getSystemAdminUser());
-        // Publish chat post
-        chatPostService.publishChatPost(jcspChatPostAccepted);
-    }
-
-    /**
      * If the job opportunity has a country associated, set that as the relocated country for the candidate and create
      * candidate note to track change.
      * @param opp Candidate Opportunity which we get the job opp and the candidate from
@@ -1003,17 +773,6 @@ public class CandidateOpportunityServiceImpl implements CandidateOpportunityServ
             candidate.setRelocatedCountry(jobOpp.getCountry());
             candidateService.save(candidate, false);
         }
-    }
-
-    /**
-     * Get candidate name and number string for automated chat posts
-     * @param candidate Candidate to get details from
-     */
-    private String constructCandidateNameNumber(Candidate candidate) {
-        // Get candidate name and number for automated chat posts
-        return candidate.getUser().getFirstName() + " "
-                + candidate.getUser().getLastName()
-                + " (" + candidate.getCandidateNumber() + ")";
     }
 
     @Transactional
